% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_design.R
\name{as.data.frame.edgar_design}
\alias{as.data.frame.edgar_design}
\title{\code{as.data.frame} method: returns the underlying \code{rows} data frame.}
\usage{
\method{as.data.frame}{edgar_design}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An \code{edgar_design} object.}

\item{row.names}{See \link[base:as.data.frame]{base::as.data.frame}. Unused.}

\item{optional}{See \link[base:as.data.frame]{base::as.data.frame}. Unused.}

\item{...}{Passed on to other methods.}
}
\value{
The \code{rows} data frame.
}
\description{
\code{as.data.frame} method: returns the underlying \code{rows} data frame.
}
