% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_latin.R
\name{design_latin}
\alias{design_latin}
\title{Generate a \code{latin} design.}
\usage{
design_latin(
  experiment_name = "",
  treatment_factor = "Variety",
  treatment_count = 3L,
  row_factor = "Row",
  row_count = 3L,
  column_factor = "Column",
  column_count = 3L,
  replicate_factor = "Square",
  treatment_names = NULL,
  seed = 0L
)
}
\arguments{
\item{experiment_name}{Optional experiment name.}

\item{treatment_factor}{Treatment column header (default "Variety").}

\item{treatment_count}{Number of treatments (>= 2).}

\item{row_factor}{Row header (default "Row").}

\item{row_count}{Number of rows. Must be a multiple of \code{treatment_count}.}

\item{column_factor}{Column header (default "Column").}

\item{column_count}{Number of columns. Must be a multiple of \code{treatment_count}.}

\item{replicate_factor}{Square header (default "Square").}

\item{treatment_names}{Optional character vector of treatment names.}

\item{seed}{Integer seed.}
}
\value{
An \code{edgar_design} S3 object.
}
\description{
Generate a \code{latin} design.
}
