% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_split_plot.R
\name{design_split_plot}
\alias{design_split_plot}
\title{Generate a \code{split_plot} design.}
\usage{
design_split_plot(
  experiment_name = "",
  block_factor = "Block",
  block_count = 2L,
  main_unit_label = "Main plot",
  main_treatment_factor = "MainTreat",
  main_treatment_count = 2L,
  sub_unit_label = "Sub-plot",
  sub_treatment_factor = "SubTreat",
  sub_treatment_count = 2L,
  main_treatment_names = NULL,
  sub_treatment_names = NULL,
  seed = 0L
)
}
\arguments{
\item{experiment_name}{Optional experiment name.}

\item{block_factor}{Block column header (default "Block").}

\item{block_count}{Number of blocks (1..100).}

\item{main_unit_label}{Whole-plot column header (default "Main plot").}

\item{main_treatment_factor}{Main treatment header (default "MainTreat").}

\item{main_treatment_count}{Number of main treatments (2..100).}

\item{sub_unit_label}{Sub-plot column header (default "Sub-plot").}

\item{sub_treatment_factor}{Sub-treatment header (default "SubTreat").}

\item{sub_treatment_count}{Number of sub-treatments (2..100).}

\item{main_treatment_names}{Optional character vector of main treatment names.}

\item{sub_treatment_names}{Optional character vector of sub-treatment names.}

\item{seed}{Integer seed.}
}
\value{
An \code{edgar_design} S3 object.
}
\description{
Generate a \code{split_plot} design.
}
