% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_two_factor_rcb.R
\name{design_two_factor_rcb}
\alias{design_two_factor_rcb}
\title{Generate a \code{two_factor_rcb} design.}
\usage{
design_two_factor_rcb(
  experiment_name = "",
  factor_a = "Treatment1",
  factor_a_count = 2L,
  factor_b = "Treatment2",
  factor_b_count = 2L,
  block_factor = "Block",
  block_count = 2L,
  unit_label = "Plot",
  factor_a_names = NULL,
  factor_b_names = NULL,
  seed = 0L
)
}
\arguments{
\item{experiment_name}{Optional experiment name.}

\item{factor_a}{Header for factor A (default "Treatment1").}

\item{factor_a_count}{Number of levels of factor A (2..100).}

\item{factor_b}{Header for factor B (default "Treatment2").}

\item{factor_b_count}{Number of levels of factor B (2..100).}

\item{block_factor}{Block column header (default "Block").}

\item{block_count}{Number of blocks (1..100).}

\item{unit_label}{Unit column header (default "Plot").}

\item{factor_a_names}{Optional character vector of factor A level names.}

\item{factor_b_names}{Optional character vector of factor B level names.}

\item{seed}{Integer seed.}
}
\value{
An \code{edgar_design} S3 object.
}
\description{
Generate a \code{two_factor_rcb} design.
}
