% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ExperimentalDesignGeneratorandRandomiser-package}
\alias{ExperimentalDesignGeneratorandRandomiser}
\alias{ExperimentalDesignGeneratorandRandomiser-package}
\title{Package-level documentation.}
\description{
EDGAR - Experimental Design Generator and Randomiser.
}
\details{
A native R implementation of EDGAR. EDGAR originated as Excel
workbooks developed by the Biometrics team at Rothamsted Research
(http://www.edgarweb.org.uk/). The algorithms were subsequently
re-implemented in the open-source Python project \code{rotsl/edgar}
(\code{rotsl/edgar}; https://rotsl.github.io/edgar/), distributed as the \code{edgar-design}
package on PyPI (https://pypi.org/project/edgar-design/). This R
package is a native R port of that Python implementation: it does
not require Python, reticulate, or any external service at runtime.

The package provides deterministic, reproducible randomisation for
nine experimental designs:
\code{cr_eq}, \code{cr_uneq}, \code{rcb}, \code{rcb_uneq}, \code{two_factor_rcb},
\code{latin}, \code{split_plot}, \code{variable_blocks}, and \code{alpha}.

The Mersenne Twister seeding and Fisher-Yates shuffle are ported
from CPython so the same integer seed produces the same design in R
as in the upstream Python implementation.

Alpha designs follow the methodology described by Patterson and
Williams (1976).
}
\references{
Patterson, H.D. & Williams, E.R. (1976). A new class of resolvable
incomplete block designs. Biometrika, 63(1), 83-92.
doi:10.1093/biomet/63.1.83

Biometrics team at Rothamsted Research, EDGAR Excel workbooks,
http://www.edgarweb.org.uk/

rotsl/edgar Python implementation, https://rotsl.github.io/edgar/

edgar-design on PyPI, https://pypi.org/project/edgar-design/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/biologyautomation/edgar-r}
  \item \url{https://rotsl.github.io/edgar/}
  \item \url{http://www.edgarweb.org.uk/}
  \item Report bugs at \url{https://github.com/biologyautomation/edgar-r/issues}
}

}
\author{
\strong{Maintainer}: BiologyAutomation \email{phonics-tiffs1i@icloud.com} (Repository owner)

Authors:
\itemize{
  \item Rohan R \email{phonics-tiffs1i@icloud.com} (\href{https://orcid.org/0009-0005-9225-1775}{ORCID}) (Author of the upstream Python implementation rotsl/edgar and this native R port) [copyright holder]
}

}
