% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_designs}
\alias{list_designs}
\title{List all registered designs with their metadata.}
\usage{
list_designs()
}
\value{
A data frame with columns \code{key}, \code{name}, \code{has_layout},
\code{has_blocks}. The full \code{info} lists are stored in the
\code{attribute }"info"`.
}
\description{
List all registered designs with their metadata.
}
\examples{
list_designs()
}
