% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{seeded_randint}
\alias{seeded_randint}
\title{Return a random integer between a and b inclusive.}
\usage{
seeded_randint(rng, a, b)
}
\arguments{
\item{rng}{An RNG returned by \code{make_rng()}.}

\item{a}{Lower bound (integer).}

\item{b}{Upper bound (integer).}
}
\value{
An integer.
}
\description{
Return a random integer between a and b inclusive.
}
