% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{seeded_sample}
\alias{seeded_sample}
\title{Return k random items from \code{items}, without replacement.}
\usage{
seeded_sample(rng, items, k)
}
\arguments{
\item{rng}{An RNG returned by \code{make_rng()}.}

\item{items}{An atomic vector.}

\item{k}{Number of items to draw.}
}
\value{
A new vector of length k.
}
\description{
Return k random items from \code{items}, without replacement.
}
