% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgar_design.R
\name{total_units}
\alias{total_units}
\title{Total number of experimental units.}
\usage{
total_units(x)
}
\arguments{
\item{x}{An \code{edgar_design} object (or any object with a \code{total_units} method).}
}
\value{
An integer.
}
\description{
Total number of experimental units.
}
